fix bug about crack
这个东西就是因为float相减导致误差累计
eg 1.0 1.9 2.8 3.7 4.6 5.5
差取整 全是0了

+++ server_core/cracker.py
@@ -179,7 +179,8 @@ class Cracker:
         ActionChains(self.driver).click_and_hold(element).perform()
         assert (track[1][0], track[1][1], track[1][2]) == (0, 0, 0)
 
-        real_track = [(x*ratio, y) for x, y, _ in track]
+        real_track = [(int(x*ratio), y) for x, y, _ in track]
+
         actions = ActionChains(self.driver)
         for i in range(2, len(real_track), 1):
             actions.move_by_offset(
